Refactor editor field wrapper
