+++ src/pages/configuration/editor/fields/SelectionFieldConfigForm.tsx
