+++ styleguide.md
@@ -4,6 +4,6 @@ Imports sorted by:
 
 1. libraries
 2. utilities/constants
-3. standard components
-4. relative components
-5. styles
+3. styles
+4. standard components
+5. relative components
